refactor validator input format helpers
